Add generated files
