+++ src/io_bank0/dormant_wake_intf.rs
@@ -1,22 +1,22 @@
-#[doc = "Register `DORMANT_WAKE_INTF1` reader"]
-pub struct R(crate::R<DORMANT_WAKE_INTF1_SPEC>);
+#[doc = "Register `DORMANT_WAKE_INTF%s` reader"]
+pub struct R(crate::R<DORMANT_WAKE_INTF_SPEC>);
 impl core::ops::Deref for R {
-    type Target = crate::R<DORMANT_WAKE_INTF1_SPEC>;
+    type Target = crate::R<DORMANT_WAKE_INTF_SPEC>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-impl From<crate::R<DORMANT_WAKE_INTF1_SPEC>> for R {
+impl From<crate::R<DORMANT_WAKE_INTF_SPEC>> for R {
     #[inline(always)]
-    fn from(reader: crate::R<DORMANT_WAKE_INTF1_SPEC>) -> Self {
+    fn from(reader: crate::R<DORMANT_WAKE_INTF_SPEC>) -> Self {
         R(reader)
     }
 }
-#[doc = "Register `DORMANT_WAKE_INTF1` writer"]
-pub struct W(crate::W<DORMANT_WAKE_INTF1_SPEC>);
+#[doc = "Register `DORMANT_WAKE_INTF%s` writer"]
+pub struct W(crate::W<DORMANT_WAKE_INTF_SPEC>);
 impl core::ops::Deref for W {
-    type Target = crate::W<DORMANT_WAKE_INTF1_SPEC>;
+    type Target = crate::W<DORMANT_WAKE_INTF_SPEC>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -28,31 +28,31 @@ impl core::ops::DerefMut for W {
         &mut self.0
     }
 }
-impl From<crate::W<DORMANT_WAKE_INTF1_SPEC>> for W {
+impl From<crate::W<DORMANT_WAKE_INTF_SPEC>> for W {
     #[inline(always)]
-    fn from(writer: crate::W<DORMANT_WAKE_INTF1_SPEC>) -> Self {
+    fn from(writer: crate::W<DORMANT_WAKE_INTF_SPEC>) -> Self {
         W(writer)
     }
 }
-#[doc = "Field `GPIO15_EDGE_HIGH` reader - "]
-pub struct GPIO15_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO15_EDGE_HIGH_R {
+#[doc = "Field `GPIO7_EDGE_HIGH` reader - "]
+pub struct GPIO7_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO7_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO15_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO7_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO15_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO7_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO15_EDGE_HIGH` writer - "]
-pub struct GPIO15_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO7_EDGE_HIGH` writer - "]
+pub struct GPIO7_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO15_EDGE_HIGH_W<'a> {
+impl<'a> GPIO7_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -70,25 +70,25 @@ impl<'a> GPIO15_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO15_EDGE_LOW` reader - "]
-pub struct GPIO15_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO15_EDGE_LOW_R {
+#[doc = "Field `GPIO7_EDGE_LOW` reader - "]
+pub struct GPIO7_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO7_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO15_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO7_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO15_EDGE_LOW_R {
+impl core::ops::Deref for GPIO7_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO15_EDGE_LOW` writer - "]
-pub struct GPIO15_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO7_EDGE_LOW` writer - "]
+pub struct GPIO7_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO15_EDGE_LOW_W<'a> {
+impl<'a> GPIO7_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -106,25 +106,25 @@ impl<'a> GPIO15_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO15_LEVEL_HIGH` reader - "]
-pub struct GPIO15_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO15_LEVEL_HIGH_R {
+#[doc = "Field `GPIO7_LEVEL_HIGH` reader - "]
+pub struct GPIO7_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO7_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO15_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO7_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO15_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO7_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO15_LEVEL_HIGH` writer - "]
-pub struct GPIO15_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO7_LEVEL_HIGH` writer - "]
+pub struct GPIO7_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO15_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO7_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -142,25 +142,25 @@ impl<'a> GPIO15_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO15_LEVEL_LOW` reader - "]
-pub struct GPIO15_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO15_LEVEL_LOW_R {
+#[doc = "Field `GPIO7_LEVEL_LOW` reader - "]
+pub struct GPIO7_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO7_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO15_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO7_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO15_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO7_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO15_LEVEL_LOW` writer - "]
-pub struct GPIO15_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO7_LEVEL_LOW` writer - "]
+pub struct GPIO7_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO15_LEVEL_LOW_W<'a> {
+impl<'a> GPIO7_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -178,25 +178,25 @@ impl<'a> GPIO15_LEVEL_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO14_EDGE_HIGH` reader - "]
-pub struct GPIO14_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO14_EDGE_HIGH_R {
+#[doc = "Field `GPIO6_EDGE_HIGH` reader - "]
+pub struct GPIO6_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO6_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO14_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO6_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO14_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO6_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO14_EDGE_HIGH` writer - "]
-pub struct GPIO14_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO6_EDGE_HIGH` writer - "]
+pub struct GPIO6_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO14_EDGE_HIGH_W<'a> {
+impl<'a> GPIO6_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -214,25 +214,25 @@ impl<'a> GPIO14_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO14_EDGE_LOW` reader - "]
-pub struct GPIO14_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO14_EDGE_LOW_R {
+#[doc = "Field `GPIO6_EDGE_LOW` reader - "]
+pub struct GPIO6_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO6_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO14_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO6_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO14_EDGE_LOW_R {
+impl core::ops::Deref for GPIO6_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO14_EDGE_LOW` writer - "]
-pub struct GPIO14_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO6_EDGE_LOW` writer - "]
+pub struct GPIO6_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO14_EDGE_LOW_W<'a> {
+impl<'a> GPIO6_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -250,25 +250,25 @@ impl<'a> GPIO14_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO14_LEVEL_HIGH` reader - "]
-pub struct GPIO14_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO14_LEVEL_HIGH_R {
+#[doc = "Field `GPIO6_LEVEL_HIGH` reader - "]
+pub struct GPIO6_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO6_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO14_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO6_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO14_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO6_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO14_LEVEL_HIGH` writer - "]
-pub struct GPIO14_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO6_LEVEL_HIGH` writer - "]
+pub struct GPIO6_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO14_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO6_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -286,25 +286,25 @@ impl<'a> GPIO14_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO14_LEVEL_LOW` reader - "]
-pub struct GPIO14_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO14_LEVEL_LOW_R {
+#[doc = "Field `GPIO6_LEVEL_LOW` reader - "]
+pub struct GPIO6_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO6_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO14_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO6_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO14_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO6_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO14_LEVEL_LOW` writer - "]
-pub struct GPIO14_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO6_LEVEL_LOW` writer - "]
+pub struct GPIO6_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO14_LEVEL_LOW_W<'a> {
+impl<'a> GPIO6_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -322,25 +322,25 @@ impl<'a> GPIO14_LEVEL_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO13_EDGE_HIGH` reader - "]
-pub struct GPIO13_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO13_EDGE_HIGH_R {
+#[doc = "Field `GPIO5_EDGE_HIGH` reader - "]
+pub struct GPIO5_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO5_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO13_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO5_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO13_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO5_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO13_EDGE_HIGH` writer - "]
-pub struct GPIO13_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO5_EDGE_HIGH` writer - "]
+pub struct GPIO5_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO13_EDGE_HIGH_W<'a> {
+impl<'a> GPIO5_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -358,25 +358,25 @@ impl<'a> GPIO13_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO13_EDGE_LOW` reader - "]
-pub struct GPIO13_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO13_EDGE_LOW_R {
+#[doc = "Field `GPIO5_EDGE_LOW` reader - "]
+pub struct GPIO5_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO5_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO13_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO5_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO13_EDGE_LOW_R {
+impl core::ops::Deref for GPIO5_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO13_EDGE_LOW` writer - "]
-pub struct GPIO13_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO5_EDGE_LOW` writer - "]
+pub struct GPIO5_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO13_EDGE_LOW_W<'a> {
+impl<'a> GPIO5_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -394,25 +394,25 @@ impl<'a> GPIO13_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO13_LEVEL_HIGH` reader - "]
-pub struct GPIO13_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO13_LEVEL_HIGH_R {
+#[doc = "Field `GPIO5_LEVEL_HIGH` reader - "]
+pub struct GPIO5_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO5_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO13_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO5_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO13_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO5_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO13_LEVEL_HIGH` writer - "]
-pub struct GPIO13_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO5_LEVEL_HIGH` writer - "]
+pub struct GPIO5_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO13_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO5_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -430,25 +430,25 @@ impl<'a> GPIO13_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO13_LEVEL_LOW` reader - "]
-pub struct GPIO13_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO13_LEVEL_LOW_R {
+#[doc = "Field `GPIO5_LEVEL_LOW` reader - "]
+pub struct GPIO5_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO5_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO13_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO5_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO13_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO5_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO13_LEVEL_LOW` writer - "]
-pub struct GPIO13_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO5_LEVEL_LOW` writer - "]
+pub struct GPIO5_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO13_LEVEL_LOW_W<'a> {
+impl<'a> GPIO5_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -466,25 +466,25 @@ impl<'a> GPIO13_LEVEL_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO12_EDGE_HIGH` reader - "]
-pub struct GPIO12_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO12_EDGE_HIGH_R {
+#[doc = "Field `GPIO4_EDGE_HIGH` reader - "]
+pub struct GPIO4_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO4_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO12_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO4_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO12_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO4_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO12_EDGE_HIGH` writer - "]
-pub struct GPIO12_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO4_EDGE_HIGH` writer - "]
+pub struct GPIO4_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO12_EDGE_HIGH_W<'a> {
+impl<'a> GPIO4_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -502,25 +502,25 @@ impl<'a> GPIO12_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO12_EDGE_LOW` reader - "]
-pub struct GPIO12_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO12_EDGE_LOW_R {
+#[doc = "Field `GPIO4_EDGE_LOW` reader - "]
+pub struct GPIO4_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO4_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO12_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO4_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO12_EDGE_LOW_R {
+impl core::ops::Deref for GPIO4_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO12_EDGE_LOW` writer - "]
-pub struct GPIO12_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO4_EDGE_LOW` writer - "]
+pub struct GPIO4_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO12_EDGE_LOW_W<'a> {
+impl<'a> GPIO4_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -538,25 +538,25 @@ impl<'a> GPIO12_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO12_LEVEL_HIGH` reader - "]
-pub struct GPIO12_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO12_LEVEL_HIGH_R {
+#[doc = "Field `GPIO4_LEVEL_HIGH` reader - "]
+pub struct GPIO4_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO4_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO12_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO4_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO12_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO4_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO12_LEVEL_HIGH` writer - "]
-pub struct GPIO12_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO4_LEVEL_HIGH` writer - "]
+pub struct GPIO4_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO12_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO4_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -574,25 +574,25 @@ impl<'a> GPIO12_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO12_LEVEL_LOW` reader - "]
-pub struct GPIO12_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO12_LEVEL_LOW_R {
+#[doc = "Field `GPIO4_LEVEL_LOW` reader - "]
+pub struct GPIO4_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO4_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO12_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO4_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO12_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO4_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO12_LEVEL_LOW` writer - "]
-pub struct GPIO12_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO4_LEVEL_LOW` writer - "]
+pub struct GPIO4_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO12_LEVEL_LOW_W<'a> {
+impl<'a> GPIO4_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -610,25 +610,25 @@ impl<'a> GPIO12_LEVEL_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO11_EDGE_HIGH` reader - "]
-pub struct GPIO11_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO11_EDGE_HIGH_R {
+#[doc = "Field `GPIO3_EDGE_HIGH` reader - "]
+pub struct GPIO3_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO3_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO11_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO3_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO11_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO3_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO11_EDGE_HIGH` writer - "]
-pub struct GPIO11_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO3_EDGE_HIGH` writer - "]
+pub struct GPIO3_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO11_EDGE_HIGH_W<'a> {
+impl<'a> GPIO3_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -646,25 +646,25 @@ impl<'a> GPIO11_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO11_EDGE_LOW` reader - "]
-pub struct GPIO11_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO11_EDGE_LOW_R {
+#[doc = "Field `GPIO3_EDGE_LOW` reader - "]
+pub struct GPIO3_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO3_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO11_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO3_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO11_EDGE_LOW_R {
+impl core::ops::Deref for GPIO3_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO11_EDGE_LOW` writer - "]
-pub struct GPIO11_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO3_EDGE_LOW` writer - "]
+pub struct GPIO3_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO11_EDGE_LOW_W<'a> {
+impl<'a> GPIO3_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -682,25 +682,25 @@ impl<'a> GPIO11_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO11_LEVEL_HIGH` reader - "]
-pub struct GPIO11_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO11_LEVEL_HIGH_R {
+#[doc = "Field `GPIO3_LEVEL_HIGH` reader - "]
+pub struct GPIO3_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO3_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO11_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO3_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO11_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO3_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO11_LEVEL_HIGH` writer - "]
-pub struct GPIO11_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO3_LEVEL_HIGH` writer - "]
+pub struct GPIO3_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO11_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO3_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -718,25 +718,25 @@ impl<'a> GPIO11_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO11_LEVEL_LOW` reader - "]
-pub struct GPIO11_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO11_LEVEL_LOW_R {
+#[doc = "Field `GPIO3_LEVEL_LOW` reader - "]
+pub struct GPIO3_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO3_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO11_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO3_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO11_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO3_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO11_LEVEL_LOW` writer - "]
-pub struct GPIO11_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO3_LEVEL_LOW` writer - "]
+pub struct GPIO3_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO11_LEVEL_LOW_W<'a> {
+impl<'a> GPIO3_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -754,25 +754,25 @@ impl<'a> GPIO11_LEVEL_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO10_EDGE_HIGH` reader - "]
-pub struct GPIO10_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO10_EDGE_HIGH_R {
+#[doc = "Field `GPIO2_EDGE_HIGH` reader - "]
+pub struct GPIO2_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO2_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO10_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO2_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO10_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO2_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO10_EDGE_HIGH` writer - "]
-pub struct GPIO10_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO2_EDGE_HIGH` writer - "]
+pub struct GPIO2_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO10_EDGE_HIGH_W<'a> {
+impl<'a> GPIO2_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -790,25 +790,25 @@ impl<'a> GPIO10_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO10_EDGE_LOW` reader - "]
-pub struct GPIO10_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO10_EDGE_LOW_R {
+#[doc = "Field `GPIO2_EDGE_LOW` reader - "]
+pub struct GPIO2_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO2_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO10_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO2_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO10_EDGE_LOW_R {
+impl core::ops::Deref for GPIO2_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO10_EDGE_LOW` writer - "]
-pub struct GPIO10_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO2_EDGE_LOW` writer - "]
+pub struct GPIO2_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO10_EDGE_LOW_W<'a> {
+impl<'a> GPIO2_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -826,25 +826,25 @@ impl<'a> GPIO10_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO10_LEVEL_HIGH` reader - "]
-pub struct GPIO10_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO10_LEVEL_HIGH_R {
+#[doc = "Field `GPIO2_LEVEL_HIGH` reader - "]
+pub struct GPIO2_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO2_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO10_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO2_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO10_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO2_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO10_LEVEL_HIGH` writer - "]
-pub struct GPIO10_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO2_LEVEL_HIGH` writer - "]
+pub struct GPIO2_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO10_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO2_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -862,25 +862,25 @@ impl<'a> GPIO10_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO10_LEVEL_LOW` reader - "]
-pub struct GPIO10_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO10_LEVEL_LOW_R {
+#[doc = "Field `GPIO2_LEVEL_LOW` reader - "]
+pub struct GPIO2_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO2_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO10_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO2_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO10_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO2_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO10_LEVEL_LOW` writer - "]
-pub struct GPIO10_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO2_LEVEL_LOW` writer - "]
+pub struct GPIO2_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO10_LEVEL_LOW_W<'a> {
+impl<'a> GPIO2_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -898,25 +898,25 @@ impl<'a> GPIO10_LEVEL_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO9_EDGE_HIGH` reader - "]
-pub struct GPIO9_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO9_EDGE_HIGH_R {
+#[doc = "Field `GPIO1_EDGE_HIGH` reader - "]
+pub struct GPIO1_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO1_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO9_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO1_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO9_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO1_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO9_EDGE_HIGH` writer - "]
-pub struct GPIO9_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO1_EDGE_HIGH` writer - "]
+pub struct GPIO1_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO9_EDGE_HIGH_W<'a> {
+impl<'a> GPIO1_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -934,25 +934,25 @@ impl<'a> GPIO9_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO9_EDGE_LOW` reader - "]
-pub struct GPIO9_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO9_EDGE_LOW_R {
+#[doc = "Field `GPIO1_EDGE_LOW` reader - "]
+pub struct GPIO1_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO1_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO9_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO1_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO9_EDGE_LOW_R {
+impl core::ops::Deref for GPIO1_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO9_EDGE_LOW` writer - "]
-pub struct GPIO9_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO1_EDGE_LOW` writer - "]
+pub struct GPIO1_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO9_EDGE_LOW_W<'a> {
+impl<'a> GPIO1_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -970,25 +970,25 @@ impl<'a> GPIO9_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO9_LEVEL_HIGH` reader - "]
-pub struct GPIO9_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO9_LEVEL_HIGH_R {
+#[doc = "Field `GPIO1_LEVEL_HIGH` reader - "]
+pub struct GPIO1_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO1_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO9_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO1_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO9_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO1_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO9_LEVEL_HIGH` writer - "]
-pub struct GPIO9_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO1_LEVEL_HIGH` writer - "]
+pub struct GPIO1_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO9_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO1_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -1006,25 +1006,25 @@ impl<'a> GPIO9_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO9_LEVEL_LOW` reader - "]
-pub struct GPIO9_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO9_LEVEL_LOW_R {
+#[doc = "Field `GPIO1_LEVEL_LOW` reader - "]
+pub struct GPIO1_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO1_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO9_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO1_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO9_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO1_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO9_LEVEL_LOW` writer - "]
-pub struct GPIO9_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO1_LEVEL_LOW` writer - "]
+pub struct GPIO1_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO9_LEVEL_LOW_W<'a> {
+impl<'a> GPIO1_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -1042,25 +1042,25 @@ impl<'a> GPIO9_LEVEL_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO8_EDGE_HIGH` reader - "]
-pub struct GPIO8_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO8_EDGE_HIGH_R {
+#[doc = "Field `GPIO0_EDGE_HIGH` reader - "]
+pub struct GPIO0_EDGE_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO0_EDGE_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO8_EDGE_HIGH_R(crate::FieldReader::new(bits))
+        GPIO0_EDGE_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO8_EDGE_HIGH_R {
+impl core::ops::Deref for GPIO0_EDGE_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO8_EDGE_HIGH` writer - "]
-pub struct GPIO8_EDGE_HIGH_W<'a> {
+#[doc = "Field `GPIO0_EDGE_HIGH` writer - "]
+pub struct GPIO0_EDGE_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO8_EDGE_HIGH_W<'a> {
+impl<'a> GPIO0_EDGE_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -1078,25 +1078,25 @@ impl<'a> GPIO8_EDGE_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO8_EDGE_LOW` reader - "]
-pub struct GPIO8_EDGE_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO8_EDGE_LOW_R {
+#[doc = "Field `GPIO0_EDGE_LOW` reader - "]
+pub struct GPIO0_EDGE_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO0_EDGE_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO8_EDGE_LOW_R(crate::FieldReader::new(bits))
+        GPIO0_EDGE_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO8_EDGE_LOW_R {
+impl core::ops::Deref for GPIO0_EDGE_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO8_EDGE_LOW` writer - "]
-pub struct GPIO8_EDGE_LOW_W<'a> {
+#[doc = "Field `GPIO0_EDGE_LOW` writer - "]
+pub struct GPIO0_EDGE_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO8_EDGE_LOW_W<'a> {
+impl<'a> GPIO0_EDGE_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -1114,25 +1114,25 @@ impl<'a> GPIO8_EDGE_LOW_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO8_LEVEL_HIGH` reader - "]
-pub struct GPIO8_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
-impl GPIO8_LEVEL_HIGH_R {
+#[doc = "Field `GPIO0_LEVEL_HIGH` reader - "]
+pub struct GPIO0_LEVEL_HIGH_R(crate::FieldReader<bool, bool>);
+impl GPIO0_LEVEL_HIGH_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO8_LEVEL_HIGH_R(crate::FieldReader::new(bits))
+        GPIO0_LEVEL_HIGH_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO8_LEVEL_HIGH_R {
+impl core::ops::Deref for GPIO0_LEVEL_HIGH_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO8_LEVEL_HIGH` writer - "]
-pub struct GPIO8_LEVEL_HIGH_W<'a> {
+#[doc = "Field `GPIO0_LEVEL_HIGH` writer - "]
+pub struct GPIO0_LEVEL_HIGH_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO8_LEVEL_HIGH_W<'a> {
+impl<'a> GPIO0_LEVEL_HIGH_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -1150,25 +1150,25 @@ impl<'a> GPIO8_LEVEL_HIGH_W<'a> {
         self.w
     }
 }
-#[doc = "Field `GPIO8_LEVEL_LOW` reader - "]
-pub struct GPIO8_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
-impl GPIO8_LEVEL_LOW_R {
+#[doc = "Field `GPIO0_LEVEL_LOW` reader - "]
+pub struct GPIO0_LEVEL_LOW_R(crate::FieldReader<bool, bool>);
+impl GPIO0_LEVEL_LOW_R {
     pub(crate) fn new(bits: bool) -> Self {
-        GPIO8_LEVEL_LOW_R(crate::FieldReader::new(bits))
+        GPIO0_LEVEL_LOW_R(crate::FieldReader::new(bits))
     }
 }
-impl core::ops::Deref for GPIO8_LEVEL_LOW_R {
+impl core::ops::Deref for GPIO0_LEVEL_LOW_R {
     type Target = crate::FieldReader<bool, bool>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-#[doc = "Field `GPIO8_LEVEL_LOW` writer - "]
-pub struct GPIO8_LEVEL_LOW_W<'a> {
+#[doc = "Field `GPIO0_LEVEL_LOW` writer - "]
+pub struct GPIO0_LEVEL_LOW_W<'a> {
     w: &'a mut W,
 }
-impl<'a> GPIO8_LEVEL_LOW_W<'a> {
+impl<'a> GPIO0_LEVEL_LOW_W<'a> {
     #[doc = r"Sets the field bit"]
     #[inline(always)]
     pub fn set_bit(self) -> &'a mut W {
@@ -1189,325 +1189,325 @@ impl<'a> GPIO8_LEVEL_LOW_W<'a> {
 impl R {
     #[doc = "Bit 31"]
     #[inline(always)]
-    pub fn gpio15_edge_high(&self) -> GPIO15_EDGE_HIGH_R {
-        GPIO15_EDGE_HIGH_R::new(((self.bits >> 31) & 0x01) != 0)
+    pub fn gpio7_edge_high(&self) -> GPIO7_EDGE_HIGH_R {
+        GPIO7_EDGE_HIGH_R::new(((self.bits >> 31) & 0x01) != 0)
     }
     #[doc = "Bit 30"]
     #[inline(always)]
-    pub fn gpio15_edge_low(&self) -> GPIO15_EDGE_LOW_R {
-        GPIO15_EDGE_LOW_R::new(((self.bits >> 30) & 0x01) != 0)
+    pub fn gpio7_edge_low(&self) -> GPIO7_EDGE_LOW_R {
+        GPIO7_EDGE_LOW_R::new(((self.bits >> 30) & 0x01) != 0)
     }
     #[doc = "Bit 29"]
     #[inline(always)]
-    pub fn gpio15_level_high(&self) -> GPIO15_LEVEL_HIGH_R {
-        GPIO15_LEVEL_HIGH_R::new(((self.bits >> 29) & 0x01) != 0)
+    pub fn gpio7_level_high(&self) -> GPIO7_LEVEL_HIGH_R {
+        GPIO7_LEVEL_HIGH_R::new(((self.bits >> 29) & 0x01) != 0)
     }
     #[doc = "Bit 28"]
     #[inline(always)]
-    pub fn gpio15_level_low(&self) -> GPIO15_LEVEL_LOW_R {
-        GPIO15_LEVEL_LOW_R::new(((self.bits >> 28) & 0x01) != 0)
+    pub fn gpio7_level_low(&self) -> GPIO7_LEVEL_LOW_R {
+        GPIO7_LEVEL_LOW_R::new(((self.bits >> 28) & 0x01) != 0)
     }
     #[doc = "Bit 27"]
     #[inline(always)]
-    pub fn gpio14_edge_high(&self) -> GPIO14_EDGE_HIGH_R {
-        GPIO14_EDGE_HIGH_R::new(((self.bits >> 27) & 0x01) != 0)
+    pub fn gpio6_edge_high(&self) -> GPIO6_EDGE_HIGH_R {
+        GPIO6_EDGE_HIGH_R::new(((self.bits >> 27) & 0x01) != 0)
     }
     #[doc = "Bit 26"]
     #[inline(always)]
-    pub fn gpio14_edge_low(&self) -> GPIO14_EDGE_LOW_R {
-        GPIO14_EDGE_LOW_R::new(((self.bits >> 26) & 0x01) != 0)
+    pub fn gpio6_edge_low(&self) -> GPIO6_EDGE_LOW_R {
+        GPIO6_EDGE_LOW_R::new(((self.bits >> 26) & 0x01) != 0)
     }
     #[doc = "Bit 25"]
     #[inline(always)]
-    pub fn gpio14_level_high(&self) -> GPIO14_LEVEL_HIGH_R {
-        GPIO14_LEVEL_HIGH_R::new(((self.bits >> 25) & 0x01) != 0)
+    pub fn gpio6_level_high(&self) -> GPIO6_LEVEL_HIGH_R {
+        GPIO6_LEVEL_HIGH_R::new(((self.bits >> 25) & 0x01) != 0)
     }
     #[doc = "Bit 24"]
     #[inline(always)]
-    pub fn gpio14_level_low(&self) -> GPIO14_LEVEL_LOW_R {
-        GPIO14_LEVEL_LOW_R::new(((self.bits >> 24) & 0x01) != 0)
+    pub fn gpio6_level_low(&self) -> GPIO6_LEVEL_LOW_R {
+        GPIO6_LEVEL_LOW_R::new(((self.bits >> 24) & 0x01) != 0)
     }
     #[doc = "Bit 23"]
     #[inline(always)]
-    pub fn gpio13_edge_high(&self) -> GPIO13_EDGE_HIGH_R {
-        GPIO13_EDGE_HIGH_R::new(((self.bits >> 23) & 0x01) != 0)
+    pub fn gpio5_edge_high(&self) -> GPIO5_EDGE_HIGH_R {
+        GPIO5_EDGE_HIGH_R::new(((self.bits >> 23) & 0x01) != 0)
     }
     #[doc = "Bit 22"]
     #[inline(always)]
-    pub fn gpio13_edge_low(&self) -> GPIO13_EDGE_LOW_R {
-        GPIO13_EDGE_LOW_R::new(((self.bits >> 22) & 0x01) != 0)
+    pub fn gpio5_edge_low(&self) -> GPIO5_EDGE_LOW_R {
+        GPIO5_EDGE_LOW_R::new(((self.bits >> 22) & 0x01) != 0)
     }
     #[doc = "Bit 21"]
     #[inline(always)]
-    pub fn gpio13_level_high(&self) -> GPIO13_LEVEL_HIGH_R {
-        GPIO13_LEVEL_HIGH_R::new(((self.bits >> 21) & 0x01) != 0)
+    pub fn gpio5_level_high(&self) -> GPIO5_LEVEL_HIGH_R {
+        GPIO5_LEVEL_HIGH_R::new(((self.bits >> 21) & 0x01) != 0)
     }
     #[doc = "Bit 20"]
     #[inline(always)]
-    pub fn gpio13_level_low(&self) -> GPIO13_LEVEL_LOW_R {
-        GPIO13_LEVEL_LOW_R::new(((self.bits >> 20) & 0x01) != 0)
+    pub fn gpio5_level_low(&self) -> GPIO5_LEVEL_LOW_R {
+        GPIO5_LEVEL_LOW_R::new(((self.bits >> 20) & 0x01) != 0)
     }
     #[doc = "Bit 19"]
     #[inline(always)]
-    pub fn gpio12_edge_high(&self) -> GPIO12_EDGE_HIGH_R {
-        GPIO12_EDGE_HIGH_R::new(((self.bits >> 19) & 0x01) != 0)
+    pub fn gpio4_edge_high(&self) -> GPIO4_EDGE_HIGH_R {
+        GPIO4_EDGE_HIGH_R::new(((self.bits >> 19) & 0x01) != 0)
     }
     #[doc = "Bit 18"]
     #[inline(always)]
-    pub fn gpio12_edge_low(&self) -> GPIO12_EDGE_LOW_R {
-        GPIO12_EDGE_LOW_R::new(((self.bits >> 18) & 0x01) != 0)
+    pub fn gpio4_edge_low(&self) -> GPIO4_EDGE_LOW_R {
+        GPIO4_EDGE_LOW_R::new(((self.bits >> 18) & 0x01) != 0)
     }
     #[doc = "Bit 17"]
     #[inline(always)]
-    pub fn gpio12_level_high(&self) -> GPIO12_LEVEL_HIGH_R {
-        GPIO12_LEVEL_HIGH_R::new(((self.bits >> 17) & 0x01) != 0)
+    pub fn gpio4_level_high(&self) -> GPIO4_LEVEL_HIGH_R {
+        GPIO4_LEVEL_HIGH_R::new(((self.bits >> 17) & 0x01) != 0)
     }
     #[doc = "Bit 16"]
     #[inline(always)]
-    pub fn gpio12_level_low(&self) -> GPIO12_LEVEL_LOW_R {
-        GPIO12_LEVEL_LOW_R::new(((self.bits >> 16) & 0x01) != 0)
+    pub fn gpio4_level_low(&self) -> GPIO4_LEVEL_LOW_R {
+        GPIO4_LEVEL_LOW_R::new(((self.bits >> 16) & 0x01) != 0)
     }
     #[doc = "Bit 15"]
     #[inline(always)]
-    pub fn gpio11_edge_high(&self) -> GPIO11_EDGE_HIGH_R {
-        GPIO11_EDGE_HIGH_R::new(((self.bits >> 15) & 0x01) != 0)
+    pub fn gpio3_edge_high(&self) -> GPIO3_EDGE_HIGH_R {
+        GPIO3_EDGE_HIGH_R::new(((self.bits >> 15) & 0x01) != 0)
     }
     #[doc = "Bit 14"]
     #[inline(always)]
-    pub fn gpio11_edge_low(&self) -> GPIO11_EDGE_LOW_R {
-        GPIO11_EDGE_LOW_R::new(((self.bits >> 14) & 0x01) != 0)
+    pub fn gpio3_edge_low(&self) -> GPIO3_EDGE_LOW_R {
+        GPIO3_EDGE_LOW_R::new(((self.bits >> 14) & 0x01) != 0)
     }
     #[doc = "Bit 13"]
     #[inline(always)]
-    pub fn gpio11_level_high(&self) -> GPIO11_LEVEL_HIGH_R {
-        GPIO11_LEVEL_HIGH_R::new(((self.bits >> 13) & 0x01) != 0)
+    pub fn gpio3_level_high(&self) -> GPIO3_LEVEL_HIGH_R {
+        GPIO3_LEVEL_HIGH_R::new(((self.bits >> 13) & 0x01) != 0)
     }
     #[doc = "Bit 12"]
     #[inline(always)]
-    pub fn gpio11_level_low(&self) -> GPIO11_LEVEL_LOW_R {
-        GPIO11_LEVEL_LOW_R::new(((self.bits >> 12) & 0x01) != 0)
+    pub fn gpio3_level_low(&self) -> GPIO3_LEVEL_LOW_R {
+        GPIO3_LEVEL_LOW_R::new(((self.bits >> 12) & 0x01) != 0)
     }
     #[doc = "Bit 11"]
     #[inline(always)]
-    pub fn gpio10_edge_high(&self) -> GPIO10_EDGE_HIGH_R {
-        GPIO10_EDGE_HIGH_R::new(((self.bits >> 11) & 0x01) != 0)
+    pub fn gpio2_edge_high(&self) -> GPIO2_EDGE_HIGH_R {
+        GPIO2_EDGE_HIGH_R::new(((self.bits >> 11) & 0x01) != 0)
     }
     #[doc = "Bit 10"]
     #[inline(always)]
-    pub fn gpio10_edge_low(&self) -> GPIO10_EDGE_LOW_R {
-        GPIO10_EDGE_LOW_R::new(((self.bits >> 10) & 0x01) != 0)
+    pub fn gpio2_edge_low(&self) -> GPIO2_EDGE_LOW_R {
+        GPIO2_EDGE_LOW_R::new(((self.bits >> 10) & 0x01) != 0)
     }
     #[doc = "Bit 9"]
     #[inline(always)]
-    pub fn gpio10_level_high(&self) -> GPIO10_LEVEL_HIGH_R {
-        GPIO10_LEVEL_HIGH_R::new(((self.bits >> 9) & 0x01) != 0)
+    pub fn gpio2_level_high(&self) -> GPIO2_LEVEL_HIGH_R {
+        GPIO2_LEVEL_HIGH_R::new(((self.bits >> 9) & 0x01) != 0)
     }
     #[doc = "Bit 8"]
     #[inline(always)]
-    pub fn gpio10_level_low(&self) -> GPIO10_LEVEL_LOW_R {
-        GPIO10_LEVEL_LOW_R::new(((self.bits >> 8) & 0x01) != 0)
+    pub fn gpio2_level_low(&self) -> GPIO2_LEVEL_LOW_R {
+        GPIO2_LEVEL_LOW_R::new(((self.bits >> 8) & 0x01) != 0)
     }
     #[doc = "Bit 7"]
     #[inline(always)]
-    pub fn gpio9_edge_high(&self) -> GPIO9_EDGE_HIGH_R {
-        GPIO9_EDGE_HIGH_R::new(((self.bits >> 7) & 0x01) != 0)
+    pub fn gpio1_edge_high(&self) -> GPIO1_EDGE_HIGH_R {
+        GPIO1_EDGE_HIGH_R::new(((self.bits >> 7) & 0x01) != 0)
     }
     #[doc = "Bit 6"]
     #[inline(always)]
-    pub fn gpio9_edge_low(&self) -> GPIO9_EDGE_LOW_R {
-        GPIO9_EDGE_LOW_R::new(((self.bits >> 6) & 0x01) != 0)
+    pub fn gpio1_edge_low(&self) -> GPIO1_EDGE_LOW_R {
+        GPIO1_EDGE_LOW_R::new(((self.bits >> 6) & 0x01) != 0)
     }
     #[doc = "Bit 5"]
     #[inline(always)]
-    pub fn gpio9_level_high(&self) -> GPIO9_LEVEL_HIGH_R {
-        GPIO9_LEVEL_HIGH_R::new(((self.bits >> 5) & 0x01) != 0)
+    pub fn gpio1_level_high(&self) -> GPIO1_LEVEL_HIGH_R {
+        GPIO1_LEVEL_HIGH_R::new(((self.bits >> 5) & 0x01) != 0)
     }
     #[doc = "Bit 4"]
     #[inline(always)]
-    pub fn gpio9_level_low(&self) -> GPIO9_LEVEL_LOW_R {
-        GPIO9_LEVEL_LOW_R::new(((self.bits >> 4) & 0x01) != 0)
+    pub fn gpio1_level_low(&self) -> GPIO1_LEVEL_LOW_R {
+        GPIO1_LEVEL_LOW_R::new(((self.bits >> 4) & 0x01) != 0)
     }
     #[doc = "Bit 3"]
     #[inline(always)]
-    pub fn gpio8_edge_high(&self) -> GPIO8_EDGE_HIGH_R {
-        GPIO8_EDGE_HIGH_R::new(((self.bits >> 3) & 0x01) != 0)
+    pub fn gpio0_edge_high(&self) -> GPIO0_EDGE_HIGH_R {
+        GPIO0_EDGE_HIGH_R::new(((self.bits >> 3) & 0x01) != 0)
     }
     #[doc = "Bit 2"]
     #[inline(always)]
-    pub fn gpio8_edge_low(&self) -> GPIO8_EDGE_LOW_R {
-        GPIO8_EDGE_LOW_R::new(((self.bits >> 2) & 0x01) != 0)
+    pub fn gpio0_edge_low(&self) -> GPIO0_EDGE_LOW_R {
+        GPIO0_EDGE_LOW_R::new(((self.bits >> 2) & 0x01) != 0)
     }
     #[doc = "Bit 1"]
     #[inline(always)]
-    pub fn gpio8_level_high(&self) -> GPIO8_LEVEL_HIGH_R {
-        GPIO8_LEVEL_HIGH_R::new(((self.bits >> 1) & 0x01) != 0)
+    pub fn gpio0_level_high(&self) -> GPIO0_LEVEL_HIGH_R {
+        GPIO0_LEVEL_HIGH_R::new(((self.bits >> 1) & 0x01) != 0)
     }
     #[doc = "Bit 0"]
     #[inline(always)]
-    pub fn gpio8_level_low(&self) -> GPIO8_LEVEL_LOW_R {
-        GPIO8_LEVEL_LOW_R::new((self.bits & 0x01) != 0)
+    pub fn gpio0_level_low(&self) -> GPIO0_LEVEL_LOW_R {
+        GPIO0_LEVEL_LOW_R::new((self.bits & 0x01) != 0)
     }
 }
 impl W {
     #[doc = "Bit 31"]
     #[inline(always)]
-    pub fn gpio15_edge_high(&mut self) -> GPIO15_EDGE_HIGH_W {
-        GPIO15_EDGE_HIGH_W { w: self }
+    pub fn gpio7_edge_high(&mut self) -> GPIO7_EDGE_HIGH_W {
+        GPIO7_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 30"]
     #[inline(always)]
-    pub fn gpio15_edge_low(&mut self) -> GPIO15_EDGE_LOW_W {
-        GPIO15_EDGE_LOW_W { w: self }
+    pub fn gpio7_edge_low(&mut self) -> GPIO7_EDGE_LOW_W {
+        GPIO7_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 29"]
     #[inline(always)]
-    pub fn gpio15_level_high(&mut self) -> GPIO15_LEVEL_HIGH_W {
-        GPIO15_LEVEL_HIGH_W { w: self }
+    pub fn gpio7_level_high(&mut self) -> GPIO7_LEVEL_HIGH_W {
+        GPIO7_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 28"]
     #[inline(always)]
-    pub fn gpio15_level_low(&mut self) -> GPIO15_LEVEL_LOW_W {
-        GPIO15_LEVEL_LOW_W { w: self }
+    pub fn gpio7_level_low(&mut self) -> GPIO7_LEVEL_LOW_W {
+        GPIO7_LEVEL_LOW_W { w: self }
     }
     #[doc = "Bit 27"]
     #[inline(always)]
-    pub fn gpio14_edge_high(&mut self) -> GPIO14_EDGE_HIGH_W {
-        GPIO14_EDGE_HIGH_W { w: self }
+    pub fn gpio6_edge_high(&mut self) -> GPIO6_EDGE_HIGH_W {
+        GPIO6_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 26"]
     #[inline(always)]
-    pub fn gpio14_edge_low(&mut self) -> GPIO14_EDGE_LOW_W {
-        GPIO14_EDGE_LOW_W { w: self }
+    pub fn gpio6_edge_low(&mut self) -> GPIO6_EDGE_LOW_W {
+        GPIO6_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 25"]
     #[inline(always)]
-    pub fn gpio14_level_high(&mut self) -> GPIO14_LEVEL_HIGH_W {
-        GPIO14_LEVEL_HIGH_W { w: self }
+    pub fn gpio6_level_high(&mut self) -> GPIO6_LEVEL_HIGH_W {
+        GPIO6_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 24"]
     #[inline(always)]
-    pub fn gpio14_level_low(&mut self) -> GPIO14_LEVEL_LOW_W {
-        GPIO14_LEVEL_LOW_W { w: self }
+    pub fn gpio6_level_low(&mut self) -> GPIO6_LEVEL_LOW_W {
+        GPIO6_LEVEL_LOW_W { w: self }
     }
     #[doc = "Bit 23"]
     #[inline(always)]
-    pub fn gpio13_edge_high(&mut self) -> GPIO13_EDGE_HIGH_W {
-        GPIO13_EDGE_HIGH_W { w: self }
+    pub fn gpio5_edge_high(&mut self) -> GPIO5_EDGE_HIGH_W {
+        GPIO5_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 22"]
     #[inline(always)]
-    pub fn gpio13_edge_low(&mut self) -> GPIO13_EDGE_LOW_W {
-        GPIO13_EDGE_LOW_W { w: self }
+    pub fn gpio5_edge_low(&mut self) -> GPIO5_EDGE_LOW_W {
+        GPIO5_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 21"]
     #[inline(always)]
-    pub fn gpio13_level_high(&mut self) -> GPIO13_LEVEL_HIGH_W {
-        GPIO13_LEVEL_HIGH_W { w: self }
+    pub fn gpio5_level_high(&mut self) -> GPIO5_LEVEL_HIGH_W {
+        GPIO5_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 20"]
     #[inline(always)]
-    pub fn gpio13_level_low(&mut self) -> GPIO13_LEVEL_LOW_W {
-        GPIO13_LEVEL_LOW_W { w: self }
+    pub fn gpio5_level_low(&mut self) -> GPIO5_LEVEL_LOW_W {
+        GPIO5_LEVEL_LOW_W { w: self }
     }
     #[doc = "Bit 19"]
     #[inline(always)]
-    pub fn gpio12_edge_high(&mut self) -> GPIO12_EDGE_HIGH_W {
-        GPIO12_EDGE_HIGH_W { w: self }
+    pub fn gpio4_edge_high(&mut self) -> GPIO4_EDGE_HIGH_W {
+        GPIO4_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 18"]
     #[inline(always)]
-    pub fn gpio12_edge_low(&mut self) -> GPIO12_EDGE_LOW_W {
-        GPIO12_EDGE_LOW_W { w: self }
+    pub fn gpio4_edge_low(&mut self) -> GPIO4_EDGE_LOW_W {
+        GPIO4_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 17"]
     #[inline(always)]
-    pub fn gpio12_level_high(&mut self) -> GPIO12_LEVEL_HIGH_W {
-        GPIO12_LEVEL_HIGH_W { w: self }
+    pub fn gpio4_level_high(&mut self) -> GPIO4_LEVEL_HIGH_W {
+        GPIO4_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 16"]
     #[inline(always)]
-    pub fn gpio12_level_low(&mut self) -> GPIO12_LEVEL_LOW_W {
-        GPIO12_LEVEL_LOW_W { w: self }
+    pub fn gpio4_level_low(&mut self) -> GPIO4_LEVEL_LOW_W {
+        GPIO4_LEVEL_LOW_W { w: self }
     }
     #[doc = "Bit 15"]
     #[inline(always)]
-    pub fn gpio11_edge_high(&mut self) -> GPIO11_EDGE_HIGH_W {
-        GPIO11_EDGE_HIGH_W { w: self }
+    pub fn gpio3_edge_high(&mut self) -> GPIO3_EDGE_HIGH_W {
+        GPIO3_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 14"]
     #[inline(always)]
-    pub fn gpio11_edge_low(&mut self) -> GPIO11_EDGE_LOW_W {
-        GPIO11_EDGE_LOW_W { w: self }
+    pub fn gpio3_edge_low(&mut self) -> GPIO3_EDGE_LOW_W {
+        GPIO3_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 13"]
     #[inline(always)]
-    pub fn gpio11_level_high(&mut self) -> GPIO11_LEVEL_HIGH_W {
-        GPIO11_LEVEL_HIGH_W { w: self }
+    pub fn gpio3_level_high(&mut self) -> GPIO3_LEVEL_HIGH_W {
+        GPIO3_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 12"]
     #[inline(always)]
-    pub fn gpio11_level_low(&mut self) -> GPIO11_LEVEL_LOW_W {
-        GPIO11_LEVEL_LOW_W { w: self }
+    pub fn gpio3_level_low(&mut self) -> GPIO3_LEVEL_LOW_W {
+        GPIO3_LEVEL_LOW_W { w: self }
     }
     #[doc = "Bit 11"]
     #[inline(always)]
-    pub fn gpio10_edge_high(&mut self) -> GPIO10_EDGE_HIGH_W {
-        GPIO10_EDGE_HIGH_W { w: self }
+    pub fn gpio2_edge_high(&mut self) -> GPIO2_EDGE_HIGH_W {
+        GPIO2_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 10"]
     #[inline(always)]
-    pub fn gpio10_edge_low(&mut self) -> GPIO10_EDGE_LOW_W {
-        GPIO10_EDGE_LOW_W { w: self }
+    pub fn gpio2_edge_low(&mut self) -> GPIO2_EDGE_LOW_W {
+        GPIO2_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 9"]
     #[inline(always)]
-    pub fn gpio10_level_high(&mut self) -> GPIO10_LEVEL_HIGH_W {
-        GPIO10_LEVEL_HIGH_W { w: self }
+    pub fn gpio2_level_high(&mut self) -> GPIO2_LEVEL_HIGH_W {
+        GPIO2_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 8"]
     #[inline(always)]
-    pub fn gpio10_level_low(&mut self) -> GPIO10_LEVEL_LOW_W {
-        GPIO10_LEVEL_LOW_W { w: self }
+    pub fn gpio2_level_low(&mut self) -> GPIO2_LEVEL_LOW_W {
+        GPIO2_LEVEL_LOW_W { w: self }
     }
     #[doc = "Bit 7"]
     #[inline(always)]
-    pub fn gpio9_edge_high(&mut self) -> GPIO9_EDGE_HIGH_W {
-        GPIO9_EDGE_HIGH_W { w: self }
+    pub fn gpio1_edge_high(&mut self) -> GPIO1_EDGE_HIGH_W {
+        GPIO1_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 6"]
     #[inline(always)]
-    pub fn gpio9_edge_low(&mut self) -> GPIO9_EDGE_LOW_W {
-        GPIO9_EDGE_LOW_W { w: self }
+    pub fn gpio1_edge_low(&mut self) -> GPIO1_EDGE_LOW_W {
+        GPIO1_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 5"]
     #[inline(always)]
-    pub fn gpio9_level_high(&mut self) -> GPIO9_LEVEL_HIGH_W {
-        GPIO9_LEVEL_HIGH_W { w: self }
+    pub fn gpio1_level_high(&mut self) -> GPIO1_LEVEL_HIGH_W {
+        GPIO1_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 4"]
     #[inline(always)]
-    pub fn gpio9_level_low(&mut self) -> GPIO9_LEVEL_LOW_W {
-        GPIO9_LEVEL_LOW_W { w: self }
+    pub fn gpio1_level_low(&mut self) -> GPIO1_LEVEL_LOW_W {
+        GPIO1_LEVEL_LOW_W { w: self }
     }
     #[doc = "Bit 3"]
     #[inline(always)]
-    pub fn gpio8_edge_high(&mut self) -> GPIO8_EDGE_HIGH_W {
-        GPIO8_EDGE_HIGH_W { w: self }
+    pub fn gpio0_edge_high(&mut self) -> GPIO0_EDGE_HIGH_W {
+        GPIO0_EDGE_HIGH_W { w: self }
     }
     #[doc = "Bit 2"]
     #[inline(always)]
-    pub fn gpio8_edge_low(&mut self) -> GPIO8_EDGE_LOW_W {
-        GPIO8_EDGE_LOW_W { w: self }
+    pub fn gpio0_edge_low(&mut self) -> GPIO0_EDGE_LOW_W {
+        GPIO0_EDGE_LOW_W { w: self }
     }
     #[doc = "Bit 1"]
     #[inline(always)]
-    pub fn gpio8_level_high(&mut self) -> GPIO8_LEVEL_HIGH_W {
-        GPIO8_LEVEL_HIGH_W { w: self }
+    pub fn gpio0_level_high(&mut self) -> GPIO0_LEVEL_HIGH_W {
+        GPIO0_LEVEL_HIGH_W { w: self }
     }
     #[doc = "Bit 0"]
     #[inline(always)]
-    pub fn gpio8_level_low(&mut self) -> GPIO8_LEVEL_LOW_W {
-        GPIO8_LEVEL_LOW_W { w: self }
+    pub fn gpio0_level_low(&mut self) -> GPIO0_LEVEL_LOW_W {
+        GPIO0_LEVEL_LOW_W { w: self }
     }
     #[doc = "Writes raw bits to the register."]
     #[inline(always)]
@@ -1520,21 +1520,21 @@ impl W {
 
 This register you can [`read`](crate::generic::Reg::read), [`write_with_zero`](crate::generic::Reg::write_with_zero), [`reset`](crate::generic::Reg::reset), [`write`](crate::generic::Reg::write), [`modify`](crate::generic::Reg::modify). See [API](https://docs.rs/svd2rust/#read--modify--write-api).  
 
-For information about available fields see [dormant_wake_intf1](index.html) module"]
-pub struct DORMANT_WAKE_INTF1_SPEC;
-impl crate::RegisterSpec for DORMANT_WAKE_INTF1_SPEC {
+For information about available fields see [dormant_wake_intf](index.html) module"]
+pub struct DORMANT_WAKE_INTF_SPEC;
+impl crate::RegisterSpec for DORMANT_WAKE_INTF_SPEC {
     type Ux = u32;
 }
-#[doc = "`read()` method returns [dormant_wake_intf1::R](R) reader structure"]
-impl crate::Readable for DORMANT_WAKE_INTF1_SPEC {
+#[doc = "`read()` method returns [dormant_wake_intf::R](R) reader structure"]
+impl crate::Readable for DORMANT_WAKE_INTF_SPEC {
     type Reader = R;
 }
-#[doc = "`write(|w| ..)` method takes [dormant_wake_intf1::W](W) writer structure"]
-impl crate::Writable for DORMANT_WAKE_INTF1_SPEC {
+#[doc = "`write(|w| ..)` method takes [dormant_wake_intf::W](W) writer structure"]
+impl crate::Writable for DORMANT_WAKE_INTF_SPEC {
     type Writer = W;
 }
-#[doc = "`reset()` method sets DORMANT_WAKE_INTF1 to value 0"]
-impl crate::Resettable for DORMANT_WAKE_INTF1_SPEC {
+#[doc = "`reset()` method sets DORMANT_WAKE_INTF%s to value 0"]
+impl crate::Resettable for DORMANT_WAKE_INTF_SPEC {
     #[inline(always)]
     fn reset_value() -> Self::Ux {
         0
