+++ src/io_bank0/dormant_wake_ints.rs
@@ -1,15 +1,15 @@
-#[doc = "Register `DORMANT_WAKE_INTS0` reader"]
-pub struct R(crate::R<DORMANT_WAKE_INTS0_SPEC>);
+#[doc = "Register `DORMANT_WAKE_INTS%s` reader"]
+pub struct R(crate::R<DORMANT_WAKE_INTS_SPEC>);
 impl core::ops::Deref for R {
-    type Target = crate::R<DORMANT_WAKE_INTS0_SPEC>;
+    type Target = crate::R<DORMANT_WAKE_INTS_SPEC>;
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
-impl From<crate::R<DORMANT_WAKE_INTS0_SPEC>> for R {
+impl From<crate::R<DORMANT_WAKE_INTS_SPEC>> for R {
     #[inline(always)]
-    fn from(reader: crate::R<DORMANT_WAKE_INTS0_SPEC>) -> Self {
+    fn from(reader: crate::R<DORMANT_WAKE_INTS_SPEC>) -> Self {
         R(reader)
     }
 }
@@ -627,17 +627,17 @@ impl R {
 
 This register you can [`read`](crate::generic::Reg::read). See [API](https://docs.rs/svd2rust/#read--modify--write-api).  
 
-For information about available fields see [dormant_wake_ints0](index.html) module"]
-pub struct DORMANT_WAKE_INTS0_SPEC;
-impl crate::RegisterSpec for DORMANT_WAKE_INTS0_SPEC {
+For information about available fields see [dormant_wake_ints](index.html) module"]
+pub struct DORMANT_WAKE_INTS_SPEC;
+impl crate::RegisterSpec for DORMANT_WAKE_INTS_SPEC {
     type Ux = u32;
 }
-#[doc = "`read()` method returns [dormant_wake_ints0::R](R) reader structure"]
-impl crate::Readable for DORMANT_WAKE_INTS0_SPEC {
+#[doc = "`read()` method returns [dormant_wake_ints::R](R) reader structure"]
+impl crate::Readable for DORMANT_WAKE_INTS_SPEC {
     type Reader = R;
 }
-#[doc = "`reset()` method sets DORMANT_WAKE_INTS0 to value 0"]
-impl crate::Resettable for DORMANT_WAKE_INTS0_SPEC {
+#[doc = "`reset()` method sets DORMANT_WAKE_INTS%s to value 0"]
+impl crate::Resettable for DORMANT_WAKE_INTS_SPEC {
     #[inline(always)]
     fn reset_value() -> Self::Ux {
         0
